requirements.txt
# Web framework
fastapi==0.135.1

# ASGI server
uvicorn[standard]==0.42.0

# Database (ORM + async support + driver)
sqlalchemy[asyncio]==2.0.46
asyncpg==0.31.0

# Migrations
alembic==1.18.4

# Data validation and settings
pydantic[email]==2.11.0
pydantic-settings==2.13.1

# Authentication
PyJWT==2.12.1
pwdlib[argon2]==0.3.0
